Broken, but runs.

+++ source/java/fds-project/fds/src/main/java/com/formationds/iodriver/operations/DeleteBucket.java
@@ -8,7 +8,7 @@ import com.amazonaws.services.s3.AmazonS3Client;
 
 import com.formationds.commons.NullArgumentException;
 import com.formationds.iodriver.endpoints.S3Endpoint;
-import com.formationds.iodriver.reporters.AbstractWorkflowEventListener;
+import com.formationds.iodriver.reporters.AbstractWorkloadEventListener;
 
 /**
  * Delete an S3 bucket.
@@ -29,12 +29,13 @@ public class DeleteBucket extends S3Operation
 
     @Override
     public void accept(S3Endpoint endpoint,
-                       AbstractWorkflowEventListener listener) throws ExecutionException
+                       AmazonS3Client client,
+                       AbstractWorkloadEventListener listener) throws ExecutionException
     {
         if (endpoint == null) throw new NullArgumentException("endpoint");
+        if (client == null) throw new NullArgumentException("client");
         if (listener == null) throw new NullArgumentException("listener");
 
-        AmazonS3Client client = getClientWrapped(endpoint);
         try
         {
             client.deleteBucket(_bucketName);
